# server/swagger.yaml
openapi: 3.0.3
info:
  title: AVO with Mie - Swagger Page
  description: |-
    This is a server for AVO with Mie website - Website for listening podcast and reading book. 
    
    Some useful links:
    - [The Pet Store repository](https://github.com/ducdatit2002/avowithmie-server)
    - [The source API definition](https://avowithmie-server.onrender.com/api-docs)

  contact:
    email: ducdatit2002@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://avowithmie-server.onrender.com/api/
  - url: http://localhost:4000/api/
...
paths:
...
  /users:
    post:
      tags:
        - Auth
      summary: Sign up a new user
      description: Create a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - gender
                - month
                - date
                - year
              properties:
                name:
                  type: string
                  example: "JohnDoe"
                email:
                  type: string
                  format: email
                  example: "johndoe@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePassword123!"
                gender:
                  type: string
                  enum:
                    - male
                    - female
                    - non-binary
                month:
                  type: string
                  example: "January"
                date:
                  type: string
                  example: "01"
                year:
                  type: string
                  example: "1990"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  userId:
                    type: string
                    example: "5f8d04034b5dc10d3684878f"
        '400':
          description: Invalid input, object invalid
        '500':
          description: Server error
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users, excluding their passwords and version key.
      parameters:
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
          description: Token for authentication
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
      security:
        - adminAuth: []
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get a user by ID
      description: Retrieve a specific user by ID, excluding their password and version key.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: User ID
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
            description: Token for authentication
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
      security:
        - auth: []
    put:
      tags:
        - Users
      summary: Update a user by ID
      description: Update the information of a specific user by ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: User ID
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
            description: Token for authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
      security:
        - auth: []
    delete:
      tags:
        - Users
      summary: Delete a user by ID
      description: Delete a specific user by ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: User ID
        - in: header
          name: x-auth-token
          required: true
          schema:
            type: string
            description: Token for administrative access
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User deleted successfully"
      security:
        - adminAuth: []
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          description: "Unique identifier for the User"
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: "Full name of the user"
          example: "John Doe"
        email:
          type: string
          format: email
          description: "Email address of the user"
          example: "johndoe@example.com"
        password:
          type: string
          format: password
          description: "Password for the user account (not shown in responses)"
          example: "SecurePassword123!"
        gender:
          type: string
          enum:
            - male
            - female
            - non-binary
          description: "Gender identity of the user"
          example: "non-binary"
        birthDate:
          type: string
          format: date
          description: "Birth date of the user"
          example: "1990-01-01"
        likedPodcasts:
          type: array
          items:
            type: string
          description: "List of podcast IDs liked by the user"
          example: ["podcastId1", "podcastId2"]
        likedBooks:
          type: array
          items:
            type: string
          description: "List of book IDs liked by the user"
          example: ["bookId1", "bookId2"]
        isAdmin:
          type: boolean
          description: "Flag indicating if the user has administrative privileges"
          example: false
      required:
        - id
        - name
        - email
        - password
        - gender
        - birthDate
...
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
      description: API key authentication
    adminApiKey:
      type: apiKey
      in: header
      name: x-auth-token
      description: API key authentication for admin endpoints